+++ test/performance/logging.cc
@@ -16,6 +16,7 @@
 */
 #include <gtest/gtest.h>
 
+#include <algorithm>
 #include <atomic>
 #include <map>
 #include <thread>